% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/genetime.R
\name{genetime}
\alias{genetime}
\title{genetime}
\usage{
genetime(genelist, thresh = 0.95, out.pdf = FALSE,
  out.name = "genetime.pdf", iters = 1000)
}
\arguments{
\item{genelist}{A list of gene symbol identifiers}

\item{thresh}{The quantile of expression values to take per gene.  Default is 0.95}

\item{out.pdf}{output pdf of temporal enrichment patterns, defaults to FALSE.  Default output name (set with out.name) is "genetime.pdf"}

\item{out.name}{name of output pdf of temporal enrichment patterns, default is "genetime.pdf"}

\item{iters}{Number of iterations for permutations to create null distributions,  Default is 1000}
}
\description{
This function will predict "important" time periods for a list of gene probes based on RNA-seq expression.
}
\examples{
mygenes = "ADNP" "AGAP2" "ANK2" "APH1A"
thresh = 0.99
genetime(genelist=mygenes,thresh=thresh,out.pdf=TRUE,out.name="/home/myfile.pdf")
}
\keyword{enrichment}
\keyword{gene}
\keyword{temporal}

