% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getColors.R
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors()
}
\description{
Returns 16 colors for brain regions in developmental enrichment package
}
\examples{
colors = getColors()
}
\keyword{brain}
\keyword{colors}
\keyword{region}

