% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(justgenes = FALSE)
}
\arguments{
\item{justgenes}{Return only gene list (and not full data). Default is FALSE}
}
\description{
Returns all RNA probes defined in developmental enrichment package
}
\examples{
getData()
}
\keyword{developmental}
\keyword{gene}
\keyword{lookup}

