% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/probeLookup.R
\name{probeLookup}
\alias{probeLookup}
\title{probeLookup}
\usage{
probeLookup(genelist)
}
\arguments{
\item{genelist}{A list of gene symbol identifiers}
}
\description{
Returns overlap of user gene list with RNA associated genes defined in developmental enrichment package
}
\examples{
probeLookup()
}
\keyword{RNA}
\keyword{developmental}
\keyword{gene}
\keyword{lookup}

